[enh #20] ライブラリのinterfaceクラスの実装
-ライブラリの機能を統括するinterfaceクラス実装
-上記に伴い、現在実装しているクラスのフォルダ整理完了
-ベクター2Dと3d構造体実装
-テスト駆動完了
