+++ LibraryProject/Application/App/App.cpp
@@ -16,12 +16,8 @@ namespace app
 /* Public Functions ------------------------------------------------------------------------------------------- */
 
 App::App(void)
-	: m_pGraphicsDevice(nullptr)
-	, m_pTextureManager(nullptr)
-	, m_pVertex2DManagaer(nullptr)
+	: m_pLibrary(sl::ISharokuLibrary::Instance())
 	, m_IsEnd(false)
-	, m_TexID(INT_MAX)
-	, m_VtxID(INT_MAX)
 {}
 
 App::~App(void)
@@ -29,40 +25,32 @@ App::~App(void)
 	Finalize();
 }
 
-void App::Initialize(sl::dx11::GraphicsDevice* pGraphicsDevice)
+void App::Initialize()
 {
-	m_pGraphicsDevice = pGraphicsDevice;
-	m_pTextureManager = new sl::dx11::TextureManager(m_pGraphicsDevice->GetDevice(), m_pGraphicsDevice->GetDeviceContext());
-	m_TexID = m_pTextureManager->LoadTexture("rinne.png");
-	//m_TexID = m_pTextureManager->LoadTexture("BLUE.png");
-
-	m_pVertex2DManagaer = new sl::dx11::Vertex2DManager();
-	if(!m_pVertex2DManagaer->Initialize(m_pGraphicsDevice))
-	{
-		return;
-	}
+	m_IDs.m_TexID = m_pLibrary->LoadTexture("rinne.png");
+	//m_IDs.m_TexID = m_pLibrary->LoadTexture("BLUE.png");
 
 	sl::fRect size = {0.0f, 0.0f, 300.f, 300.f};
 	sl::fRect uv = {0.0f, 0.0f, 1.0f, 1.0f};
 
-	m_VtxID = m_pVertex2DManagaer->CreateVertex(size, uv);
+	m_IDs.m_VtxID= m_pLibrary->CreateVertex2D(size, uv);
 }
 
 void App::Finalize(void)
 {
-	delete m_pVertex2DManagaer;
-	delete m_pTextureManager;
+	m_pLibrary->ReleaseVertexALL();
+	m_pLibrary->ReleaseTexALL();
 }
 
 bool App::Update(void)
 {
 	Control();
 
-	m_pGraphicsDevice->StartRender();
+	m_pLibrary->StartRender();
 
 	Draw();
 
-	m_pGraphicsDevice->EndRender();
+	m_pLibrary->EndRender();
 
 	return m_IsEnd;
 }
@@ -74,16 +62,7 @@ void App::Control(void)
 
 void App::Draw(void)
 {
-	m_pVertex2DManagaer->SetupShadar();
-	m_pVertex2DManagaer->SetupConstantBuffer(D3DXVECTOR2(10.0f, 10.0f));
-	m_pVertex2DManagaer->SetupBlendState();
-	m_pTextureManager->SetUpTexture(m_TexID);
-
-	UINT stride = sizeof(sl::dx11::BasicVertex);
-	UINT offset = 0;
-	ID3D11Buffer* pVertexBuffer = m_pVertex2DManagaer->GetBuffer(m_VtxID);
-	m_pGraphicsDevice->GetDeviceContext()->IASetVertexBuffers(0, 1, &pVertexBuffer, &stride, &offset);
-	m_pGraphicsDevice->GetDeviceContext()->Draw(4, 0);
+	m_pLibrary->Draw2D(m_IDs, sl::SLVECTOR2(0.0f,0.0f));
 }
 
 }	// namespace app
