+++ LibraryProject/SharokuLibrary/sl/Library/Vertex/dx11Vertex/dx11VertexManager.h
@@ -1,0 +1,31 @@
+﻿//==================================================================================================================================//
+//!< @file		dx11VertexManager.h
+//!< @brief		dx11::VertexManagerクラスヘッダ
+//!< @author	T.Haga
+//==================================================================================================================================//
+
+#ifndef DX11_VERTEX_MANAGER_H
+#define DX11_VERTEX_MANAGER_H
+
+namespace sl
+{
+namespace dx11
+{
+
+class VertexManager
+{
+
+public:
+	VertexManager();
+	~VertexManager();
+
+};
+
+}	// namespace dx11
+}	// namespace sl
+
+#endif	// DX11_VERTEX_MANAGER_H
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
