--- LibraryProject/SharokuLibrary/sl/Texture/dx11Texture/dx11Texture.cpp
@@ -1,31 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11Texture.cpp
-//!< @brief		dx11::Textureクラス実装
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-/* Includes --------------------------------------------------------------------------------------------------- */
-
-#include "dx11Texture.h"
-
-/* Namespace -------------------------------------------------------------------------------------------------- */
-
-namespace sl
-{
-namespace dx11
-{
-
-/* Public Functions ------------------------------------------------------------------------------------------- */
-
-Texture::Texture()
-{}
-
-Texture::~Texture(void)
-{}
-
-}	// namespace dx11
-}	// namespace sl
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
