--- LibraryProject/SharokuLibrary/sl/Texture/dx11Texture/dx11Texture.h
@@ -1,40 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11Texture.h
-//!< @brief		dx11::Textureクラスヘッダ
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-#ifndef SL_DX11_TEXTURE_H
-#define SL_DX11_TEXTURE_H
-
-namespace sl
-{
-namespace dx11
-{
-
-//======================================================================//
-//!< DirectX11のTextureのクラス
-//======================================================================//
-class Texture
-{
-
-public:
-	/** Constructor */
-	Texture(void);
-
-	/** Destructor */
-	~Texture(void);
-
-private:
-
-};
-
-}	// namespace dx11
-}	// namespace sl
-
-#endif	// SL_DX11_TEXTURE_H
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
-
