--- LibraryProject/SharokuLibrary/sl/Vertex/dx11Vertex/dx11Vertex.cpp
@@ -1,27 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11Vertex.cpp
-//!< @brief		dx11::Vertexクラス実装
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-#include "dx11Vertex.h"
-
-namespace sl
-{
-namespace dx11
-{
-
-Vertex::Vertex()
-{
-}
-
-Vertex::~Vertex()
-{
-}
-
-}	// namespace dx11
-}	// namespace sl
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
