--- LibraryProject/SharokuLibrary/sl/Vertex/dx11Vertex/dx11Vertex.h
@@ -1,29 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11Vertex.h
-//!< @brief		dx11::Vertexクラスヘッダ
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-#ifndef DX11_VERTEX_H
-#define DX11_VERTEX_H
-
-namespace sl
-{
-namespace dx11
-{
-
-class Vertex
-{
-public:
-	Vertex();
-	~Vertex();
-};
-
-}	// namespace dx11
-}	// namespace sl
-
-#endif // DX11_VERTEX_H
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
