--- LibraryProject/SharokuLibrary/sl/Vertex/dx11Vertex/dx11VertexManager.cpp
@@ -1,25 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11VertexManager.cpp
-//!< @brief		dx11::VertexManagerクラス実装
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-#include "dx11VertexManager.h"
-
-namespace sl
-{
-namespace dx11
-{
-
-VertexManager::VertexManager()
-{}
-
-VertexManager::~VertexManager()
-{}
-
-}	// namespace dx11
-}	// namespace sl
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
